+++ src/main/comp34120/ex2/regression/Regression.java
@@ -4,7 +4,7 @@ import comp34120.ex2.Record;
 
 public interface Regression {
 
-    float predictFollowerPrice(int day, double leaderPrice);
+    float predictFollowerPrice(double day, double leaderPrice);
 
     void update(Record record);
